requirements.txt
Flask==2.3.2
matplotlib
pandas==1.3.4
gunicorn==20.1.0
spotipy==2.23.0
python-dotenv==1.0.0

